data/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-storage-blob>=12.27.1",
    "dotenv>=0.9.9",
    "datasets>=2.21.0",
]
